+++ packages/client/src/components/game-end/props.tsx
