+++ packages/client/src/hooks/index.ts
@@ -1,1 +1,2 @@
-export * from './useUser'
+export * from './use-full-screen'
+export * from './use-user'
