+++ packages/client/src/hooks/use-full-screen.ts
