+++ packages/client/src/pages/about-game/about-game.tsx
