+++ packages/client/src/pages/about-game/index.ts
@@ -1,0 +1,1 @@
+export * from './about-game'
