--- packages/client/src/pages/aboutGame/index.ts
@@ -1,1 +1,0 @@
-export * from './AboutGame'
