+++ packages/client/src/pages/index.ts
@@ -1,10 +1,10 @@
-export * from './aboutGame'
-export * from './leaderboard'
-export * from './singIn'
-export * from './singUp'
-export * from './userProfile'
+export * from './about-game'
+export * from './leader-board'
+export * from './sing-in'
+export * from './sing-up'
+export * from './user-profile'
 export * from './settings'
-export * from './Page404'
-export * from './Game'
-export * from './Page500'
-export * from './GameMenu'
+export * from './page-404'
+export * from './_game'
+export * from './page-500'
+export * from './game-menu'
