+++ packages/client/src/pages/leader-board/index.ts
@@ -1,0 +1,1 @@
+export * from './leader-board'
