--- packages/client/src/pages/leaderboard/index.ts
@@ -1,1 +1,0 @@
-export * from './leaderboard'
