+++ packages/client/src/pages/sing-up/index.ts
@@ -1,0 +1,1 @@
+export * from './sing-up'
