--- packages/client/src/pages/singIn/index.ts
@@ -1,1 +1,0 @@
-export * from './SingIn'
