--- packages/client/src/pages/singUp/index.ts
@@ -1,1 +1,0 @@
-export * from './SingUp'
